Feature: sharing event
With these changes a user can share an event directly with her friends'
cozy: in the share pop-up she only has to select the Cozy button
(symbolized by a cloud) to switch from mail to cozy.

For this button to appear she must have entered beforehand the url of
the cozy in the contact information: select the option "Cozy" in the
drop-down list "Add field" in the "Additional Informations" section.